renamed dof_utilities.h to dof_utilities.hpp
