+++ applications/GeoMechanicsApplication/custom_conditions/U_Pw_condition.cpp
@@ -14,7 +14,7 @@
 
 // Application includes
 #include "custom_conditions/U_Pw_condition.h"
-#include "custom_utilities/dof_utilities.h"
+#include "custom_utilities/dof_utilities.hpp"
 
 namespace Kratos
 {
