+++ applications/GeoMechanicsApplication/custom_elements/Pw_element.h
@@ -20,7 +20,7 @@
 #include "custom_retention/retention_law_factory.h"
 #include "custom_utilities/check_utilities.hpp"
 #include "custom_utilities/constitutive_law_utilities.h"
-#include "custom_utilities/dof_utilities.h"
+#include "custom_utilities/dof_utilities.hpp"
 #include "custom_utilities/element_utilities.hpp"
 #include "custom_utilities/hydraulic_discharge.h"
 #include "custom_utilities/transport_equation_utilities.hpp"
