+++ applications/GeoMechanicsApplication/custom_elements/U_Pw_base_element.cpp
@@ -14,7 +14,7 @@
 // Application includes
 #include "custom_elements/U_Pw_base_element.hpp"
 #include "custom_utilities/check_utilities.hpp"
-#include "custom_utilities/dof_utilities.h"
+#include "custom_utilities/dof_utilities.hpp"
 #include "custom_utilities/equation_of_motion_utilities.h"
 #include "includes/serializer.h"
 #include "utilities/geometry_utilities.h"
