+++ applications/GeoMechanicsApplication/custom_elements/interface_element.cpp
@@ -13,7 +13,7 @@
 #include "interface_element.h"
 
 #include "custom_utilities/constitutive_law_utilities.h"
-#include "custom_utilities/dof_utilities.h"
+#include "custom_utilities/dof_utilities.hpp"
 #include "custom_utilities/element_utilities.hpp"
 #include "custom_utilities/equation_of_motion_utilities.h"
 #include "custom_utilities/extrapolation_utilities.h"
