+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_utilities/test_dof_utilities.cpp
@@ -17,7 +17,7 @@
 #include "containers/model.h"
 #include "custom_elements/U_Pw_small_strain_element.hpp"
 #include "custom_elements/plane_strain_stress_state.h"
-#include "custom_utilities/dof_utilities.h"
+#include "custom_utilities/dof_utilities.hpp"
 #include "geo_aliases.h"
 #include "geo_mechanics_application_variables.h"
 #include "includes/element.h"
